requirements.txt
python-binance==1.0.29
pybit==5.11.0
python-dotenv==1.1.1
pandas==2.3.1
numpy==2.3.2
websocket-client==1.8.0
aiohttp==3.12.15
requests==2.32.4